fix:PR comments - move triggerCriteria to detailsInfo, replace pageData with selectedItem
