+++ src/components/Alerts/Alerts.js
@@ -23,6 +23,7 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import AlertsView from './AlertsView'
 
+import { ALERTS_PAGE } from '../../constants'
 import { createAlertRowData } from '../../utils/createAlertsContent'
 import { getAlertsFiltersConfig, parseAlertsQueryParamsCallback } from './alerts.util'
 import { generatePageData } from './alerts.util'
@@ -100,7 +101,7 @@ const Alerts = () => {
     if (tableContent.length > 0) {
       const alert = tableContent.find(({ data }) => data.id && data.id === params.alertId)
       if (alert) {
-        setSelectedAlert({ ...alert.data })
+        setSelectedAlert({ ...alert.data, page: ALERTS_PAGE })
       } else {
         return setSelectedAlert({})
       }
