+++ src/components/Alerts/alerts.util.js
@@ -94,9 +94,7 @@ export const generatePageData = (handleFetchJobLogs, selectedAlert) => {
       entityType: selectedAlert.entity_kind,
       infoHeaders: alertsHeaders(selectedAlert.entity_kind),
       menu: [],
-      notifications,
-      refreshLogs: handleFetchJobLogs,
-      triggerCriteria: triggerCriteria
+      refreshLogs: handleFetchJobLogs
     },
     selectedAlert
   }
@@ -212,24 +210,3 @@ export const alertsHeaders = type => {
 
   return []
 }
-
-const triggerCriteria = criteria => {
-  if (criteria) {
-    return [
-      {
-        label: 'Trigger criteria count',
-        id: 'triggerCriteriaCount',
-        value: criteria?.count
-      },
-      {
-        label: 'Trigger criteria time period',
-        id: 'triggerCriteriaTimePeriod',
-        value: criteria?.period
-      }
-    ]
-  }
-
-  return []
-}
-
-export const notifications = [{ label: 'Notifications', id: 'notifications' }]
